+++ component/src/test/java/arez/component/RepositoryTest.java
@@ -33,15 +33,15 @@ public class RepositoryTest
     assertFalse( context.safeAction( () -> repository.contains( entity4 ) ) );
     assertFalse( context.safeAction( () -> repository.contains( entity5 ) ) );
 
-    context.safeAction( () -> repository.registerEntity( entity1 ) );
+    context.safeAction( () -> repository.attach( entity1 ) );
 
     assertTrue( context.safeAction( () -> repository.contains( entity1 ) ) );
     assertFalse( context.safeAction( () -> repository.contains( entity2 ) ) );
 
-    context.safeAction( () -> repository.registerEntity( entity2 ) );
-    context.safeAction( () -> repository.registerEntity( entity3 ) );
-    context.safeAction( () -> repository.registerEntity( entity4 ) );
-    context.safeAction( () -> repository.registerEntity( entity5 ) );
+    context.safeAction( () -> repository.attach( entity2 ) );
+    context.safeAction( () -> repository.attach( entity3 ) );
+    context.safeAction( () -> repository.attach( entity4 ) );
+    context.safeAction( () -> repository.attach( entity5 ) );
 
     assertTrue( context.safeAction( () -> repository.contains( entity1 ) ) );
     assertTrue( context.safeAction( () -> repository.contains( entity2 ) ) );
@@ -96,8 +96,8 @@ public class RepositoryTest
     final MyEntity entity2 = new MyEntity( 302 );
     final MyRepository repository = MyRepository.create();
 
-    context.safeAction( () -> repository.registerEntity( entity1 ) );
-    context.safeAction( () -> repository.registerEntity( entity2 ) );
+    context.safeAction( () -> repository.attach( entity1 ) );
+    context.safeAction( () -> repository.attach( entity2 ) );
 
     assertTrue( context.safeAction( () -> repository.contains( entity1 ) ) );
     assertTrue( context.safeAction( () -> repository.contains( entity2 ) ) );
@@ -117,12 +117,12 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 304 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 304 ) ) );
     final MyEntity entity = new MyEntity( 305 );
     entity.dispose();
-    context.safeAction( () -> repository.registerEntity( entity ) );
+    context.safeAction( () -> repository.attach( entity ) );
 
     final Set<Integer> ids =
       context.safeAction( () -> repository.findAll().stream().map( e -> e.getArezId() ).collect( Collectors.toSet() ) );
@@ -138,12 +138,12 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 304 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 304 ) ) );
     final MyEntity entity = new MyEntity( 305 );
     entity.dispose();
-    context.safeAction( () -> repository.registerEntity( entity ) );
+    context.safeAction( () -> repository.attach( entity ) );
 
     final int[] ids =
       context.safeAction( () -> repository.findAll( Comparator.comparing( e -> -e.getArezId() ) )
@@ -160,12 +160,12 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 304 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 304 ) ) );
     final MyEntity entity = new MyEntity( 305 );
     entity.dispose();
-    context.safeAction( () -> repository.registerEntity( entity ) );
+    context.safeAction( () -> repository.attach( entity ) );
 
     final Set<Integer> ids =
       context.safeAction( () -> repository.findAllByQuery( e -> e.getArezId() <= 303 )
@@ -181,14 +181,14 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 300 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 301 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 304 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 300 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 301 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 304 ) ) );
     final MyEntity entity = new MyEntity( 305 );
     entity.dispose();
-    context.safeAction( () -> repository.registerEntity( entity ) );
+    context.safeAction( () -> repository.attach( entity ) );
 
     final int[] ids =
       context.safeAction( () -> repository.findAllByQuery( e -> e.getArezId() <= 303,
@@ -207,8 +207,8 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
 
     final MyEntity entity = context.safeAction( () -> repository.findByQuery( e -> e.getArezId() == 302 ) );
     assertNotNull( entity );
@@ -227,8 +227,8 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
 
     final MyEntity entity = context.safeAction( () -> repository.getByQuery( e -> e.getArezId() == 302 ) );
     assertNotNull( entity );
@@ -247,8 +247,8 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
 
     final MyEntity entity = context.safeAction( () -> repository.findByArezId( 302 ) );
     assertNotNull( entity );
@@ -267,8 +267,8 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
 
     final MyEntity entity = context.safeAction( () -> repository.getByArezId( 302 ) );
     assertNotNull( entity );
@@ -296,8 +296,8 @@ public class RepositoryTest
     final ArezContext context = Arez.context();
 
     final MyRepository repository = MyRepository.create();
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 302 ) ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 302 ) ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
 
     final int[] ids = context.safeAction( () -> repository.entities().mapToInt( e -> e.getArezId() ).toArray() );
     assertEquals( ids.length, 2 );
@@ -313,8 +313,8 @@ public class RepositoryTest
     final MyRepository repository = MyRepository.create();
     final MyEntity entity = new MyEntity( 302 );
     Disposable.dispose( entity );
-    context.safeAction( () -> repository.registerEntity( entity ) );
-    context.safeAction( () -> repository.registerEntity( new MyEntity( 303 ) ) );
+    context.safeAction( () -> repository.attach( entity ) );
+    context.safeAction( () -> repository.attach( new MyEntity( 303 ) ) );
 
     final int[] ids = context.safeAction( () -> repository.entities().mapToInt( e -> e.getArezId() ).toArray() );
     assertEquals( ids.length, 1 );