+++ processor/src/test/resources/expected/com/example/repository/InjectEnabledRepositoryRepository.java
@@ -28,7 +28,7 @@ public abstract class InjectEnabledRepositoryRepository extends AbstractReposito
   @Nonnull
   public InjectEnabledRepository create(@Nonnull final String name) {
     final Arez_InjectEnabledRepository entity = new Arez_InjectEnabledRepository(name);
-    registerEntity( entity );
+    attach( entity );
     return entity;
   }